all scripts: Replace use of sqlite with subcommand
Introducing 'sfp-query' which will do the heavy lifting of accessing the
database in a predictable, maintainable, extensible, and sane fashion.

Expect that all other subcommands and extensions, present and future, to
use sfp-query and not touch the sqlite directly.

Remove database argument out of the global config, due to all of this
Remove other scripts usage of that variable, too

Closes #5